promoted flatten_params() and unflatten_params() to main fitting module
